super duper awesome 3d rrt in rviz is done
things look very nice. The algorithm works perfect and fast. In the
future I think it would be cool to add some obstacles for it to try and
navigate, and make it so when a node in the tree has a clear path to the
goal, that we are allowed to just see that and jump  to the goal across
any distance.